Cleanup in plugins: cdripper, compressor, delayaudio
Removed out commented parts of source
Implemented framenum, posnum, samplenum. A lot of cosmetics.
Added icons to configuration windows

+++ plugins/delayaudio/delayaudio.h
@@ -29,25 +29,20 @@
 #include "vframe.inc"
 #include <vector>
 
-
-
 class DelayAudio;
 class DelayAudioWindow;
 class DelayAudioTextBox;
-
-
 
 
 class DelayAudioConfig
 {
 public:
 	DelayAudioConfig();
-	
+
 	int equivalent(DelayAudioConfig &that);
 	void copy_from(DelayAudioConfig &that);
 	double length;
 };
-
 
 
 class DelayAudioThread : public Thread
@@ -55,14 +50,13 @@ class DelayAudioThread : public Thread
 public:
 	DelayAudioThread(DelayAudio *plugin);
 	~DelayAudioThread();
-	
+
 	void run();
-	
+
 	Mutex completion;
 	DelayAudioWindow *window;
 	DelayAudio *plugin;
 };
-
 
 
 class DelayAudioWindow : public BC_Window
@@ -70,7 +64,7 @@ class DelayAudioWindow : public BC_Window
 public:
 	DelayAudioWindow(DelayAudio *plugin, int x, int y);
 	~DelayAudioWindow();
-	
+
 	int create_objects();
 	int close_event();
 	void update_gui();
@@ -80,20 +74,16 @@ public:
 };
 
 
-
-
 class DelayAudioTextBox : public BC_TextBox
 {
 public:
 	DelayAudioTextBox(DelayAudio *plugin, int x, int y);
 	~DelayAudioTextBox();
-	
+
 	int handle_event();
-	
+
 	DelayAudio *plugin;
 };
-
-
 
 
 class DelayAudio : public PluginAClient
@@ -109,27 +99,18 @@ public:
 	int save_defaults();
 	void read_data(KeyFrame *keyframe);
 	void save_data(KeyFrame *keyframe);
-	int process_realtime(int64_t size, double *input_ptr, double *output_ptr);
+	int process_realtime(int size, double *input_ptr, double *output_ptr);
 	int show_gui();
 	void raise_window();
 	int set_string();
 
-
-
-
-
 	void load_configuration();
 	void update_gui();
-
-
 
 	std::vector<double> buffer;
 	DelayAudioThread *thread;
 	BC_Hash *defaults;
 	DelayAudioConfig config;
 };
-
-
-
 
 #endif
